- Fix compiler warnings for Solaris - Use of PAM_MODULE_UNKNOWN broke pam_opensc for Sun based   PAM implementations, fixed.
git-svn-id: https://www.opensc-project.org/svnp/opensc/trunk@376 c6295689-39f2-0310-b995-f0e70906c6a9

+++ src/tools/pkcs15-init.c
@@ -1890,7 +1890,7 @@ read_options_file(const char *filename)
 			if (o->has_arg != no_argument) {
 				optarg = strtok(NULL, "");
 				if (optarg) {
-					while (isspace(*optarg))
+					while (isspace((int) *optarg))
 						optarg++;
 					optarg = strdup(optarg);
 				}
